FitRandomForest {ModelDataComp}	R Documentation

Fit a random forest to a data set

Description

-

Usage

FitRandomForest(x, y, nodes = 4, ntree = 500, mtry = NULL, xvars2 = NULL, 
    ...)

Arguments

x
predictor variables

y
dependent variable

nodes
number of nodes for parallel compuation of random forest

ntree
number of trees

mtry
Number of variables randomly sampled as candidates at each split. If NULL codetuneRF will be used to determine an optimal value for mtry

xvars2

...
further arguments

Details

No details.

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

References

No reference.

See Also

FitDataModel

[Package ModelDataComp version 1.0 Index]