ModelDataComp-package {ModelDataComp}	R Documentation

Model-Data Comparison

Description

Metrics and plots for model-data comparison. The package provides functions to 1) compute objective functions (model performance metrics), to 2) fit empirical models to data (currently random forest and multiple logistic functions), and to 3) make model-data comparison plots (barplots and scatterplots of objective functions, Taylor diagrams, WollMilchSauPlot).

Details

The DESCRIPTION file: This package was not yet installed at build time.

Index: This package was not yet installed at build time.

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

[Package ModelDataComp version 1.0 Index]