RandomLHSPar {ModelDataComp}	R Documentation

Latin-hypercube sampling of parameters

Description

RandomLHSPar samples parameters based on random Latin-Hypercube sampling within given parameter ranges

Usage

RandomLHSPar(nexp, lower, upper, fixed = rep(FALSE, length(lower)))

Arguments

nexp
number of samples

lower
numeric vector of lower boundary values for each parameter

upper
numeric vector of upper borundary values for each parameter

fixed
boolean vector: should a parameter kept fixed (TRUE) or should all parameters be sampled (FALSE). If TRUE than the parameter value from the vector lower will be returned.

Value

The function returns a data.frame with sampled parameters.

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

Examples

lower <- c(0, 100, -50, 1000)
names(lower) <- letters[1:4]
upper <- c(10, 500, 50, 10000)
RandomLHSPar(10, lower, upper)

[Package ModelDataComp version 1.0 Index]