TaylorDiagram {ModelDataComp}	R Documentation

Taylor diagram

Description

Plot a Taylor diagram. This is a modification of the function taylor.diagram in the plotrix package.

Usage

TaylorDiagram(ref, model, groups = rep(1, length(model)), col = NULL, 
    plot.combined = FALSE, normalize = TRUE, sd.arcs = TRUE, 
    text.groups = NULL, text.ref = "Obs", text.combined = "All", 
    pos.cor = TRUE, ...)

Arguments

ref
numeric vector - reference values

model
numeric vector - predicted model values

groups
numeric vector that split ref and model in different groups, e.g. to indicate different models or subsets of the data

col
color for each group, should have the same length as grouping elements

plot.combined
plot also the combined set, i.e. without splitting by groups?

normalize
whether to normalize the models so that the reference has a standard deviation of 1

sd.arcs
whether to display arcs along the standard deviation axes

text.groups
text labels for the groups

text.ref
text label for the reference

text.combined
text label for the combined model/data set without grouping

pos.cor
whether to display only positive (TRUE) or all values of correlation (FALSE). If NULL, this will depend on the correlations.

...
further arguments as in taylor.diagram

Details

No details.

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

References

No reference.

See Also

BAmodel2IntegrationData

Examples

ref <- rnorm(50, 0, 5)
model1 <- ref + c(rnorm(25, 1, 2), rnorm(25, 4, 0.2))
model2 <- ref + c(rnorm(25, -5, 5), rnorm(25, 4, 0.2))
model3 <- ref + c(rnorm(25, 10, 10), rnorm(25, 4, 0.2))
data <- data.frame(ref=rep(ref, 3), model=c(model1, model2, model3), 
   group.models=rep(c("model1", "model2", "model3"), each=50), 
   group.models.subsets=rep(c("model1.subset1", "model1.subset2", "model2.subset1", "model2.subset2", "model3.subset1", "model3.subset2"), each=25))

TaylorDiagram(data$ref, data$model, data$group.models)
TaylorDiagram(data$ref, data$model, data$group.models.subsets)

[Package ModelDataComp version 1.0 Index]